Move bookmark button to a separate file
